add this in /usr/local/lib/python2.7/dist-packages/django/db/models/fields/__init__.py
class PositiveBigIntegerField(BigIntegerField):
    empty_strings_allowed = False
    description = _("Big (8 byte) positive integer")

    def db_type(self, connection):
        """
        Returns MySQL-specific column data type. Make additional checks
        to support other backends.
        """
        return 'bigint UNSIGNED'

    def formfield(self, **kwargs):
        defaults = {'min_value': 0,
                    'max_value': BigIntegerField.MAX_BIGINT * 2 - 1}
        defaults.update(kwargs)
        return super(PositiveBigIntegerField, self).formfield(**defaults)

+++ vendorData/models.py
@@ -58,7 +58,7 @@ class Slot(models.Model):
 	slot_sport_id = models.ForeignKey(Sport)
 	slot_date = models.DateField()
 	slot_duration = models.PositiveSmallIntegerField()
-	slot_availability = models.IntegerField()
+	slot_availability = models.PositiveBigIntegerField()
 
 class Event(models.Model):
 	event_club_id = models.ForeignKey(Club)
